feat: log aggregation v13, tracing v14, dashboard/reporting v13
Log Aggregation v13:
- Database migration 421: log_entries_v13 and alert_rules_v10 tables
- Alert rules v10 with pattern matching
- Files: log_aggregation/mod.rs, types.rs, store.rs

Distributed Tracing v14:
- Database migration 422: trace_sampling_rules_v13 and service_dependencies_v10 tables
- Service dependency tracking v10
- Latency analysis and capacity planning v14
- Files: distributed_tracing_v2/mod.rs, types.rs, store.rs

Dashboard & Reporting v13:
- Database migration 423: dashboard_shares_v10 and report_schedules_v11 tables
- Dashboard sharing v10
- Report scheduling v11
- Files: dashboard_reporting/mod.rs, types.rs, store.rs

Files changed: 8
- crates/civit-core/src/log_aggregation/mod.rs, types.rs, store.rs
- crates/civit-core/src/distributed_tracing_v2/mod.rs, types.rs, store.rs
- crates/civit-core/src/dashboard_reporting/mod.rs, types.rs, store.rs
- crates/civit-db/src/migrations/421_add_log_aggregation_v13.sql
- crates/civit-db/src/migrations/422_add_distributed_tracing_v14.sql
- crates/civit-db/src/migrations/423_add_dashboard_reporting_v13.sql
- crates/civit-db/src/migrations/mod.rs
- crates/civit-core/src/automation_rules.rs (fix type mismatch)

+++ crates/civit-core/src/automation_rules.rs
@@ -5005,7 +5005,7 @@ impl AutomationRuleService {
             rule_id: row.id,
             total_runs: row.run_count as i64,
             successful_runs: ((row.success_rate / 100.0) * row.run_count as f64) as i64,
-            failed_runs: row.run_count - ((row.success_rate / 100.0) * row.run_count as f64) as i64,
+            failed_runs: row.run_count as i64 - ((row.success_rate / 100.0) * row.run_count as f64) as i64,
             average_execution_time_ms: row.avg_execution_time_ms as f64,
             last_execution_time_ms: None,
             success_rate: row.success_rate,
